% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceratio.R
\name{spaceratio}
\alias{spaceratio}
\title{Function for plotting Space Ratio Charts or deriving a data frame of summary statistics.}
\usage{
spaceratio(
  density_var,
  existing_density = NULL,
  permissible_density = NULL,
  space_ratio = NULL,
  output = "plot"
)
}
\arguments{
\item{density_var}{A vector of density variables}

\item{existing_density}{A vector containing the existing density values per density variable in density_var. Input only if space_ratio not inputted}

\item{permissible_density}{A vector containing the permissible density values per density variable in density_var. Input only if space_ratio not inputted}

\item{space_ratio}{A vector of Space Ratio values per density variable in density_var. Input only if existing_density and permissible_density not inputted}

\item{output}{Output of function. Input 'plot' for the Space Ratio Chart. Input 'stats' for a data frame of summary statistics. Defaults to 'plot'}
}
\value{
A Space Ratio Chart or a data frame of summary stats.
}
\description{
Function for plotting Space Ratio Charts or deriving a data frame of summary statistics.
}
\examples{
# Density variables
d_var <- c("FAR", "DPH", "H", "GSI")

# Space Ratio values
s_val <- c(0.73, 1, 0.81, 0.91)

# Space Ratio Chart
spaceratio(density_var = d_var, space_ratio = s_val, output = 'plot')

# Summary Stats
spaceratio(density_var = d_var, space_ratio = s_val, output = 'stats')
}
